# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Affair:
    properties:
      auditId:
        type: integer
      cost:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      department:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      state:
        type: boolean
      theme:
        type: string
      updatedAt:
        type: string
    type: object
  model.Apply:
    properties:
      access:
        type: string
      affairsId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      fileUrl:
        type: string
      id:
        type: integer
      personId:
        type: integer
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  model.Feedback:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      replyContent:
        type: string
      responder:
        type: string
      state:
        type: boolean
      type:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
...
  model.PageForm:
    properties:
      name:
        type: string
      pageNum:
        type: integer
      pageSize:
        type: integer
    required:
    - pageNum
    - pageSize
    type: object
...
host: localhost:9090
info:
  contact:
    email: grezkl@protonmail.com
    name: grezkl
    url: http://github.com/grezkl
  description: e-gov server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: e-gov server
  version: "1.0"
paths:
  /affairPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 政务名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 政务管理分页
      tags:
      - Affair
  /applyPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 政务申请名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 政务申请分页
      tags:
      - Apply
  /delAffairBatch/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 政务编号数组
        in: body
        name: id
        required: true
        schema:
          items:
            type: integer
          type: array
      responses: {}
      summary: 删除政务（批量）
      tags:
      - Affair
  /delApplyBatch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 政务申请编号数组
        in: path
        items:
          type: integer
        name: ids
        required: true
        type: array
      responses: {}
      summary: 撤销政务申请（批量）
      tags:
      - Apply
  /delFeedbackBatch/:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 反馈编号数组
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 撤销反馈（批量）
      tags:
      - Feedback
...
  /deleteAffairById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 政务编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 删除政务（单选）
      tags:
      - Affair
  /deleteApplyById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 政务申请编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 撤销政务申请（单个）
      tags:
      - Apply
  /deleteFeedbackById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 反馈编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 撤销反馈（单个）
      tags:
      - Feedback
...
  /feedbackPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 反馈名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 反馈分页
      tags:
      - Feedback
...
  /pullAffairByAuditId/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 审批人员编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 获取审批人员负责的政务
      tags:
      - Affair
  /pullAffairById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 政务编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据政务编号获取信息
      tags:
      - Affair
...
  /pullAffairPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 政务部门
        in: query
        name: department
        type: string
      - description: 政务主题
        in: query
        name: theme
        type: string
      responses: {}
      summary: 政务办理分页
      tags:
      - Affair
...
  /pullFeedbackByUserId/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据用户编号查询反馈
      tags:
      - Feedback
...
  /pullMyApply:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页面规格及关键词
        in: body
        name: page_form
        required: true
        schema:
          $ref: '#/definitions/model.PageForm'
      responses: {}
      summary: 获取我的政务分页
      tags:
      - Apply
...
  /pullUserById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 政务申请编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据编号查找政务申请
      tags:
      - Account
...
  /saveAffair:
    post:
      consumes:
      - application/json
      parameters:
      - description: 政务内容
        in: body
        name: affair
        required: true
        schema:
          $ref: '#/definitions/model.Affair'
      responses: {}
      summary: 政务内容保存
      tags:
      - Affair
  /saveApply:
    post:
      consumes:
      - application/json
      parameters:
      - description: 政务申请
        in: body
        name: apply
        required: true
        schema:
          $ref: '#/definitions/model.Apply'
      responses: {}
      summary: 政务申请办理
      tags:
      - Apply
  /saveFeedback:
    post:
      consumes:
      - application/json
      parameters:
      - description: 反馈内容
        in: body
        name: feedback
        required: true
        schema:
          $ref: '#/definitions/model.Feedback'
      responses: {}
      summary: 反馈提交
      tags:
      - Feedback
...
  /updateAffair:
    post:
      consumes:
      - application/json
      parameters:
      - description: 政务内容
        in: body
        name: affair
        required: true
        schema:
          $ref: '#/definitions/model.Affair'
      responses: {}
      summary: 政务内容更新
      tags:
      - Affair
...
swagger: "2.0"